drivers: add generic build of SVE files
For SVE, as for AVX2 and AVX-512, support building files for these ISAs
from the top-level drivers/meson.build file, rather than having each
driver re-implement it.

This removes the remaining build task for drivers in DPDK which is
being done by a driver itself, rather than in the generic drivers'
build rules.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>