net/mlx5: fix counter pool init error propagation
In case of an error in mlx5_hws_cnt_pool_create(),
value stored in ret is used as "error code".
There are however a few cases inside this function,
when this variable is not set, leading to failed
assertions (e.g., when requested number of counters is bigger
than maximum supported).

This patch addresses these cases, by propagating the rte_errno
set by failing functions.

Fixes: e1c83d295dd9 ("net/mlx5: support ASO actions with non-template flow")
Cc: mkashani@nvidia.com
Cc: stable@dpdk.org

Signed-off-by: Dariusz Sosnowski <dsosnowski@nvidia.com>
Acked-by: Bing Zhao <bingz@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>