eal: add getline() function for Windows
Existing DPDK code uses getline(), which is a POSIX function and is
not available in the Windows APIs.

Instead of rewriting it or coming up with some other replacement, this
patch makes use of the implementation provided by NetBSD to make it
possible to compile code dependent on getline on Windows.

Signed-off-by: Andre Muezerie <andremue@linux.microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>