eal: add asprintf() function for Windows
The asprintf function is not part of the C standard library but is a
GNU extension commonly available in Unix-like systems. It dynamically
allocates memory to store the formatted output string, similar to
sprintf, but avoids buffer overflow issues by automatically sizing
the buffer.

Instead of rewriting it or coming up with some other replacement, this
patch makes use of the implementation provided by Neved4.

Signed-off-by: Andre Muezerie <andremue@linux.microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>