eal/linux: unregister alarm callback before free ptr
This was flagged by Address sanitizer as a use after free. The
intr_handle ptr is shared between the main thread and the interrupt
thread, and the interrupt thread can dereference the ptr after free
is called when the main thread cleans up (from the alarm callback).

The interrupt thread never terminates (eal_intr_thread_main) so
use rte_intr_callback_unregister_sync during cleanup to
ensure the callback is removed before freeing the ptr.

To be more defensive clear out the pointer and registration
variable if we can unregister.

Bugzilla ID: 1683

Signed-off-by: Rui Ferreira <rui.ferreira1@h-partners.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>