eventdev: introduce event vector adapter
The event vector adapter offloads creation of event vectors
i.e., aggregation of 8B objects (mbufs/ptrs/u64s) into a
rte_event_vector from the CPU thereby reducing the overhead
of allocating the vector and maintaining vector state.

Applications can create a vector adapter associated with
an event queue and enqueue objects to be vectorized.
When the vector reaches the configured size or when the timeout
is reached, the vector adapter will enqueue the vector to the
event queue.

Signed-off-by: Pavan Nikhilesh <pbhagavatula@marvell.com>
Series-acked-by: Jerin Jacob <jerinj@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>