devtools/linux-uapi: conform to kernel headers install
The Linux kernel has an existing script to sanitize headers which
uses more steps to clean up the header file. Make the DPDK header
import script use the same regular expression. The kernel script
is GPL-2.0 so ok.

Not necessary for current vduse.h header only but useful if more
headers are added later. Want the distro and DPDK header to
be exact match.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>