buildtools: allow a different minimum meson version for Windows
There is a minimum meson version specified in the DPDK meson project
section, which has been documented. This string is parsed by
buildtools\get-min-meson-version.py and this information is used by
lab automation to install the corresponding meson package on the
build machine.

Turns out that the meson version specified on the DPDK project
(0.57.x) is buggy on Windows: it has issues related to path
manipulation, and these issues are causing failures. Therefore,
a newer meson is required on Windows.

To avoid bringing the minimum requirements up for all operating
systems before a more appropriate release, this patch implements
a mechanism allowing a different version to be specified for Windows.

Signed-off-by: Andre Muezerie <andremue@linux.microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>