mempool: record mempool objects operations history
This feature is designed to monitor the lifecycle of
mempool objects as they move between the application
and the PMD.

It will allow us to track the operations and transitions
of each mempool object throughout the system, helping in
debugging and understanding objects flow.

added a bitmap to the mempool object header that records
the most recent operations.
Each operation is represented by a 4-bit value, and for each mempool
object the history field stores 16 steps.

Signed-off-by: Shani Peretz <shperetz@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>