doc: use core lists rather than masks in guides
For the guides for applications and examples, use the "-l" flag rather
than the "-c" flag to pass in core lists rather than core masks. While
for the trivial examples given, core masks work fine and may be easier;
in real-world cases, with potentially hundreds of cores, core lists are
a much better fix, so update the docs to reflect this.

Similarly replace instances of "-s <service-coremask>" with
"-S <service-core-list>".

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>